--- config.py
@@ -1,4 +1,0 @@
-from groq import Groq
-
-
-GROQ_API_KEY = "API"
