+++ pyproject.toml
@@ -16,3 +16,7 @@ pylint = "^3.2.6"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+
+[tool.poetry.scripts]
+prosdebugger = "prosdebugger.cli:main"
